feat, Common Targets
- Detach Target APP
- Automatic creation of Common Targets
