+++ project_manager/backend/settings.py
@@ -60,6 +60,7 @@ INSTALLED_APPS = [
     'rest_framework',
     'oauth2_provider',
     'corsheaders',
+    'targets',
     'tango',
 ]
