+++ project_manager/tango/viewsDataset.py
@@ -22,8 +22,6 @@ from PIL import Image
 import cv2
 import base64
 
-from .models import Target
-
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 root_path = os.path.dirname(os.path.dirname(BASE_DIR))
